+++ packages/avatar/src/index.ts
@@ -1,1 +1,2 @@
+import './Avatar.dark.scss';
 export {default} from './Avatar.vue';