+++ packages/forms/src/index.ts
@@ -1,4 +1,6 @@
 import './forms.scss';
+import './checkbox/VCheckbox.dark.scss';
+
 export {default as VTextarea} from './textarea/Textarea.vue';
 export {default as VCheckbox} from './checkbox/VCheckbox.vue';
 export {default as VFileUpload} from './file-input/VFileUpload.vue';
@@ -14,5 +16,6 @@ export {default as VInputGroup} from './input/VInputGroup.vue';
 export {default as VInputRange} from './input/VInputRange.vue';
 export {default as VRadio} from './radio/VRadio.vue';
 export {default as VRadioGroup} from './radio/VRadioGroup.vue';
+export * from './composables';
 export * from './form-select/types';
 export * from './file-input/types';