+++ packages/multi-select/src/index.ts
@@ -1,4 +1,5 @@
 import VMultiSelect from './VMultiSelect.vue';
+import './VMultiSelect.dark.scss';
 
 export {VMultiSelect};
 export default VMultiSelect;