+++ packages/nuxt/playground/nuxt.config.ts
@@ -4,7 +4,6 @@ export default defineNuxtConfig({
   modules: [GitsUIModule, '@nuxtjs/tailwindcss'],
   ssr: false,
   gitsUi: {
-    css: true,
-    sass: false,
+    darkMode: true,
   },
 });