+++ packages/table/src/index.ts
@@ -1,5 +1,7 @@
 import VDataTable from './VDataTable.vue';
 import VDataTablePagination from './VDataTablePagination.vue';
+import './VDataTable.dark.scss';
+import './VDataTablePagination.dark.scss';
 
 export * from './types';
 export {VDataTable, VDataTablePagination};